[RENAME] rename component
